# Hoist loads and sink store in hot loop