@stevecheckoway 2022-03-29
By the way, one reason I'm a little skeptical about the performance gain is that the common case is going to be ASCII characters (things like tag names). Thus the body of the loop is going to execute exactly once. In the worst case, it'll execute 4 times per code point, so we're only saving a handful of loads and stores.