@sebpop 2022-03-29
Nice to see that the patch is getting merged.
The patch for gumbo-parser was to improve performance of a benchmark in Geekbench.
I forgot the details, and I don't have access to Geekbench anymore after I left Samsung.