@flavorjones 2022-04-09
I'm going to spend a bit of time writing a benchmark for this to understand the impact a bit better. Sorry for the delay on responding/reviewing!