@flavorjones 2022-08-29
In a real-world document parsing benchmark there doesn't seem to be any change in performance.

Here's the benchmark, html4 parsing is presented simply for comparison (note that HTML4 parsing is not affected by the compiler option changes).

```ruby
#! /usr/bin/env ruby
# coding: utf-8

require "bundler/inline"

gemfile do
  source "https://rubygems.org"
  gem "nokogiri", path: "."
  gem "benchmark-ips"
end

require "nokogiri"
require "benchmark/ips"

input = File.read("test/files/tlm.html")
puts "input #{input.length} bytes"

html4_doc = Nokogiri::HTML4::Document.parse(input)
html5_doc = Nokogiri::HTML5::Document.parse(input)

puts RUBY_DESCRIPTION

Benchmark.ips do |x|
  x.time = 10
  x.report("html5 parse") do
    Nokogiri::HTML5::Document.parse(input)
  end
  x.report("html4 parse") do
    Nokogiri::HTML4::Document.parse(input)
  end
  x.compare!
end
```

(note that's parsing [this test file](https://github.com/sparklemotion/nokogiri/blob/main/test/files/tlm.html))

Here's the benchmark results:


_main with default compiler options:_

```
input 70095 bytes
ruby 3.1.2p20 (2022-04-12 revision 4491bb740a) [x86_64-linux]
Warming up --------------------------------------
         html5 parse    11.000  i/100ms
         html4 parse    34.000  i/100ms
Calculating -------------------------------------
         html5 parse    125.659  (±17.5%) i/s -      1.210k in  10.013864s
         html4 parse    380.538  (±22.6%) i/s -      3.638k in  10.011955s

Comparison:
         html4 parse:      380.5 i/s
         html5 parse:      125.7 i/s - 3.03x  (± 0.00) slower

```


_with this patch and default compiler options:_

```
input 70095 bytes
ruby 3.1.2p20 (2022-04-12 revision 4491bb740a) [x86_64-linux]
Warming up --------------------------------------
         html5 parse    11.000  i/100ms
         html4 parse    33.000  i/100ms
Calculating -------------------------------------
         html5 parse    120.569  (±18.2%) i/s -      1.166k in  10.077918s
         html4 parse    365.071  (±25.5%) i/s -      3.465k in  10.094851s

Comparison:
         html4 parse:      365.1 i/s
         html5 parse:      120.6 i/s - 3.03x  (± 0.00) slower

```

---

_main with -O2 (see #2369):_

```
input 70095 bytes
ruby 3.1.2p20 (2022-04-12 revision 4491bb740a) [x86_64-linux]
Warming up --------------------------------------
         html5 parse    22.000  i/100ms
         html4 parse    36.000  i/100ms
Calculating -------------------------------------
         html5 parse    249.689  (±19.2%) i/s -      2.398k in  10.001083s
         html4 parse    378.405  (±20.6%) i/s -      3.672k in  10.065013s

Comparison:
         html4 parse:      378.4 i/s
         html5 parse:      249.7 i/s - 1.52x  (± 0.00) slower
```

_this patch with -O2 (see #2369):_

```
input 70095 bytes
ruby 3.1.2p20 (2022-04-12 revision 4491bb740a) [x86_64-linux]
Warming up --------------------------------------
         html5 parse    19.000  i/100ms
         html4 parse    38.000  i/100ms
Calculating -------------------------------------
         html5 parse    247.586  (±22.2%) i/s -      2.356k in  10.081576s
         html4 parse    375.071  (±18.1%) i/s -      3.648k in  10.012846s

Comparison:
         html4 parse:      375.1 i/s
         html5 parse:      247.6 i/s - 1.51x  (± 0.00) slower
```

@sebpop or @stevecheckoway any thoughts? I'd be happy to run a more specific benchmark if you can write one.