@flavorjones 2022-10-16
Bumping this - any ideas for a benchmark that will demonstrate this speedup?